+++ TwinPushSDK/ViewControllers/Inbox/TPAliasInboxViewController.h
@@ -1,0 +1,14 @@
+//
+//  TPAliasInboxViewController.h
+//  TwinPushSDK
+//
+//  Created by Guillermo Gutiérrez on 11/8/16.
+//  Copyright © 2016 TwinCoders. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "TPNotificationsInboxViewController.h"
+
+@interface TPAliasInboxViewController : TPNotificationsInboxViewController
+
+@end