+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- No changes
+- Included alias inbox functionality
 
 ## [2.0.2] - 2016-04-13
 ### Changed